+++ src/backends/x11.rs
@@ -388,6 +388,14 @@ pub struct BufferImpl<'a, D: ?Sized, W: ?Sized>(&'a mut X11Impl<D, W>);
 impl<D: HasDisplayHandle + ?Sized, W: HasWindowHandle + ?Sized> BufferInterface
     for BufferImpl<'_, D, W>
 {
+    fn width(&self) -> usize {
+        self.0.size.unwrap().0.get() as usize
+    }
+
+    fn height(&self) -> usize {
+        self.0.size.unwrap().1.get() as usize
+    }
+
     #[inline]
     fn pixels(&self) -> &[u32] {
         // SAFETY: We called `finish_wait` on the buffer, so it is safe to call `buffer()`.