+++ src/lib.rs
@@ -205,6 +205,28 @@ pub struct Buffer<'a, D, W> {
 }
 
 impl<D: HasDisplayHandle, W: HasWindowHandle> Buffer<'_, D, W> {
+    /// The amount of pixels wide the buffer is.
+    pub fn width(&self) -> usize {
+        let width = self.buffer_impl.width();
+        debug_assert_eq!(
+            width * self.buffer_impl.height(),
+            self.len(),
+            "buffer must be sized correctly"
+        );
+        width
+    }
+
+    /// The amount of pixels tall the buffer is.
+    pub fn height(&self) -> usize {
+        let height = self.buffer_impl.height();
+        debug_assert_eq!(
+            height * self.buffer_impl.width(),
+            self.len(),
+            "buffer must be sized correctly"
+        );
+        height
+    }
+
     /// `age` is the number of frames ago this buffer was last presented. So if the value is
     /// `1`, it is the same as the last frame, and if it is `2`, it is the same as the frame
     /// before that (for backends using double buffering). If the value is `0`, it is a new